backend/requirements.txt
fastapi
uvicorn
google-genai
google-cloud-tasks
google-cloud-storage
google-auth
ffmpeg-python
pydantic
python-multipart
python-dotenv
